+++ 有好货页面/id_1908022/demo/package.json
@@ -16,6 +16,8 @@
     "@babel/preset-env": "^7.6.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-transform-react-jsx": "^6.24.1",
+    "file-loader": "^4.2.0",
+    "url-loader": "^2.2.0",
     "webpack": "^4.39.3"
   },
   "dependencies": {
